requirements.txt
requests==2.32.5
PyMySQL==1.1.2
python-dotenv==1.2.2
cryptography==46.0.5
urllib3==2.6.3
selenium==4.41.0
playwright==1.58.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-dbapi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-pymysql==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
lingua-language-detector==1.3.4
langdetect==1.0.9

#playwright install chromium